examples: migrate triangle-led-front to React controls host + canvas-scoped input
