+++ apps/docs/examples/triangle-led-front/index.tsx
@@ -1,0 +1,36 @@
+'use client';
+
+import { useEffect, useRef, useState } from 'react';
+import type { ExampleRenderer } from '../../lib/example-renderer';
+import { Controls } from './controls';
+import { createRenderer } from './renderer';
+import { DEFAULT_TRIANGLE_LED_CONTROLS, type TriangleLedControls } from './types';
+
+export function Example() {
+  const canvasRef = useRef<HTMLCanvasElement>(null);
+  const rendererRef = useRef<ExampleRenderer<TriangleLedControls> | null>(null);
+  const [controls, setControls] = useState<TriangleLedControls>(DEFAULT_TRIANGLE_LED_CONTROLS);
+
+  useEffect(() => {
+    const canvas = canvasRef.current;
+    if (!canvas) return;
+    const renderer = createRenderer({ canvas, initialControls: controls });
+    rendererRef.current = renderer;
+    void renderer.ready.catch(() => {});
+    return () => {
+      rendererRef.current = null;
+      renderer.dispose();
+    };
+    // Initial controls are passed only at mount; subsequent changes use setControls.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  useEffect(() => rendererRef.current?.setControls?.(controls), [controls]);
+
+  return (
+    <div className="relative h-full w-full overflow-hidden bg-black">
+      <canvas ref={canvasRef} className="block h-full w-full touch-none" />
+      <Controls value={controls} onChange={setControls} />
+    </div>
+  );
+}
